// src/typings/payout/amount.ts
/*
 * The version of the OpenAPI document: v68
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit this class manually.
 */


export class Amount {
    /**
    * The three-character [ISO currency code](https://docs.adyen.com/development-resources/currency-codes#currency-codes).
    */
    'currency': string;
    /**
    * The amount of the transaction, in [minor units](https://docs.adyen.com/development-resources/currency-codes#minor-units).
    */
    'value': number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "currency",
            "baseName": "currency",
            "type": "string"
        },
        {
            "name": "value",
            "baseName": "value",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return Amount.attributeTypeMap;
    }
}